+++ includes/c_server.hpp
@@ -17,7 +17,7 @@ class   server {
 		std::string					_error_pages;
 		std::string					_index;
 		bool						_autoindex;
-		int							_client_body_size;
+		long int					_client_body_size;
 		std::map<std::string, std::string>	_cgi;
 		std::vector<location>		_location;
 
@@ -59,7 +59,7 @@ class   server {
 		std::string					const &getError() const;
 		std::string					const &getIndex() const;
 		bool						const &getAutoindex() const;
-		int							const &getClientBodySize() const;
+		long int					const &getClientBodySize() const;
 		std::vector<location>		const &getLocation() const;
 		std::map<std::string, std::string>	const &getCgi() const;
